COMPASS-648 reorganize test structure (#755)
* COMPASS-648 reorganize test structure

* bump hadron-build@6.0.0 for new test structure.

* try sorting test files for travis.

* test new hadron-build before bumping version.

* bump hadron-build@6.0.1 (--sort flag)

#retry https://travis-ci.com/10gen/compass/jobs/66208746
