+++ package.json
@@ -71,7 +71,7 @@
   "scripts": {
     "start": "hadron-build develop",
     "pretest": "mongodb-runner install && mongodb-runner start --port 27018",
-    "test": "npm run test-main && npm run test-renderer && hadron-build test",
+    "test": "hadron-build test",
     "posttest": "mongodb-runner stop --port 27018",
     "check": "mongodb-js-precommit ./src/app/*.js ./src/app/**/**/*.js ./src/{app/**/*.js,main/**/*.js} ./test/*.js ./src/internal-packages/**/*.js ./src/internal-packages/**/**/*{.js,.jsx} ./src/internal-packages/**/**/**/*{.js,.jsx} ./src/internal-packages/**/**/**/**/*{.js,.jsx}",
     "ci": "npm run check && npm run test",
@@ -79,11 +79,12 @@
     "compile-ui": "hadron-build ui",
     "fmt": "mongodb-js-fmt ./*.js src/{**/*.js,*.js} test/{**/*.js,*.js}",
     "release": "hadron-build release",
-    "test-functional": "npm test -- --functional",
-    "test-main": "electron-mocha ./tests/main-process/*.test.js",
-    "test-packages": "npm test -- --packages",
-    "test-renderer": "electron-mocha --renderer ./tests/renderer-process/*.test.js",
     "test-unit": "npm test -- --unit",
+    "test-enzyme": "npm test -- --enzyme",
+    "test-packages": "npm test -- --packages",
+    "test-main": "npm test -- --main",
+    "test-renderer": "npm test -- --renderer",
+    "test-functional": "npm test -- --functional",
     "test-release": "npm test -- --release",
     "upload": "hadron-build upload",
     "evergreen-expansions": "hadron-build info --format=yaml --flatten > expansions.yml"
@@ -109,7 +110,7 @@
     "ampersand-view": "^9.0.0",
     "ampersand-view-switcher": "^2.0.0",
     "app-migrations": "mongodb-js/app-migrations#v0.1.2",
-    "async": "^1.5.0",
+    "async": "^1.5.2",
     "babel-register": "^6.7.2",
     "backoff": "^2.4.1",
     "bootstrap": "https://github.com/twbs/bootstrap/archive/v3.3.5.tar.gz",
@@ -191,7 +192,7 @@
     "electron-prebuilt": "1.2.8",
     "enzyme": "^2.5.1",
     "eslint-config-mongodb-js": "^2.2.0",
-    "hadron-build": "^5.1.9",
+    "hadron-build": "^6.0.1",
     "jsdom": "^9.8.3",
     "mocha": "^3.1.2",
     "mock-require": "^2.0.1",
