+++ test/enzyme/sampling-message-component.test.js
@@ -5,30 +5,23 @@ const chaiEnzyme = require('chai-enzyme');
 const expect = chai.expect;
 const React = require('react');
 const sinon = require('sinon');
-const {mount, shallow} = require('enzyme');
+const {shallow} = require('enzyme');
 const AppRegistry = require('hadron-app-registry');
 
 chai.use(chaiEnzyme());
 
-describe('<RuleBuilder />', () => {
+describe('<SamplingMessage />', () => {
   const appRegistry = app.appRegistry;
   const appInstance = app.instance;
-
-  const template = {
-    editState: 'unmodified',
-    validationAction: 'warn',
-    validationLevel: 'moderate',
-    validationRules: [],
-    serverVersion: '3.4.0',
-    validate: function() {}
-  };
 
   beforeEach(() => {
     // Mock the AppRegistry with a new one so tests don't complain about
     // appRegistry.getComponent (i.e. appRegistry being undefined)
     app.appRegistry = new AppRegistry();
+    app.appRegistry.registerAction('CRUD.Actions', sinon.spy());
+    app.appRegistry.registerStore('CRUD.ResetDocumentListStore', sinon.spy());
 
-    this.RuleBuilder = require('../src/internal-packages/validation/lib/components/rule-builder');
+    this.SamplingMessage = require('../../src/internal-packages/query/lib/component/sampling-message');
   });
   afterEach(() => {
     // Restore properties on the global app object,
@@ -39,24 +32,22 @@ describe('<RuleBuilder />', () => {
 
   context('when collection is not writable', () => {
     beforeEach(() => {
-      const props = Object.assign({isWritable: false}, template);
-      this.component = mount(<this.RuleBuilder {... props} />);
+      this.component = shallow(<this.SamplingMessage isWritable={false} insertHandler={() => {}} />);
     });
 
-    it('disables the ADD RULE button', () => {
-      const state = this.component.find('.btn.btn-xs.btn-success');
+    it('disables the INSERT DOCUMENT button', () => {
+      const state = this.component.find('.btn.btn-primary.btn-xs.open-insert');
       expect(state).to.be.disabled();
     });
   });
 
-  context('when collection is not writable', () => {
+  context('when collection is writable', () => {
     beforeEach(() => {
-      const props = Object.assign({isWritable: true}, template);
-      this.component = mount(<this.RuleBuilder {... props} />);
+      this.component = shallow(<this.SamplingMessage isWritable={true} insertHandler={() => {}} />);
     });
 
-    it('disables the ADD RULE button', () => {
-      const state = this.component.find('.btn.btn-xs.btn-success');
+    it('disables the INSERT DOCUMENT button', () => {
+      const state = this.component.find('.btn.btn-primary.btn-xs.open-insert');
       expect(state).to.not.be.disabled();
     });
   });
