+++ test/enzyme/sortable-table.test.js
@@ -8,7 +8,7 @@ const sinon = require('sinon');
 const shallow = require('enzyme').shallow;
 const mount = require('enzyme').mount;
 const bootstrap = require('react-bootstrap');
-const SortableTable = require('../src/internal-packages/app/lib/components/sortable-table');
+const SortableTable = require('../../src/internal-packages/app/lib/components/sortable-table');
 
 chai.use(chaiEnzyme());
