+++ test/renderer/zoom.test.js
