+++ test/unit/keytar.test.js
